modified: replace new tag with existing same tag
