+++ srcs/ai/saving/structure/utils/__init__.py
@@ -3,4 +3,4 @@ from ai.saving.structure.utils.memo_and_tags_converter import convert_memos_and_
 from ai.saving.structure.utils.tag_converter import convert_tag
 from ai.saving.structure.utils.relations_converter import convert_relations
 from ai.saving.structure.utils.metadata_extractor.metadata_extractor import extract_and_assign_metadata
-from ai.saving.structure.utils.get_tag_name_to_id_dict import get_tag_name_to_id_dict
+from ai.saving.structure.utils.get_tag_name_to_id_dict import get_tag_dict
