+++ srcs/ai/saving/structure/utils/directory_format.py
@@ -1,26 +1,14 @@
 from fastapi import HTTPException
 from ai.saving.utils import get_tag_relations_from_db
-from ai.database.collections.tag_store import tag_collection, TAG_UID_NAME, TAG_ID_NAME, TAG_CONTENT_NAME, TAG_ROOT_NAME
+from ai.database.collections.tag_store import TAG_ROOT_NAME
 
 
-def get_formatted_directories(user_id: str) -> str:
-    tag_id_to_name, tag_name_to_id=_get_tags_from_db(user_id)
+def get_formatted_directories(user_id: str, tag_id_to_name: dict[str, str], tag_name_to_id: dict[str, str]) -> str:
     graph: dict[str, list[str]]=get_tag_relations_from_db(user_id)
     if TAG_ROOT_NAME not in tag_name_to_id:
         raise HTTPException(status_code=500, headers={"/memo/structures": "root tag not found (@)"})
     
     return _format_graph(graph, tag_id_to_name, tag_name_to_id[TAG_ROOT_NAME])
-
-def _get_tags_from_db(user_id: str) -> tuple[dict[str, str], dict[str, str]]:
-    id_to_name: dict[str, str]={}
-    name_to_id: dict[str, str]={}
-    raw_tags=tag_collection.find({TAG_UID_NAME: user_id})
-    
-    for raw_tag in raw_tags:
-        id_to_name[raw_tag[TAG_ID_NAME]]=raw_tag[TAG_CONTENT_NAME]
-        name_to_id[raw_tag[TAG_CONTENT_NAME]]=raw_tag[TAG_ID_NAME]
-    
-    return id_to_name, name_to_id
     
 def _format_graph(graph: dict[str, list[str]], tag_id_to_name: dict[str, str], now_id: str, depth: int=1) -> str:
     result: str=f'{tag_id_to_name[now_id]}\n'
